requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
apipkg==1.5
astor==0.8.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.4.0
cryptography==3.1.1
ddt==1.4.1
execnet==1.7.1
flake8==3.8.3
idna==2.10
iniconfig==1.0.1
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.5.0
packaging==20.4
parameterized==0.7.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.0.2
pytest-ast-transformer==1.0.3
pytest-forked==1.3.0
pytest-ordering==0.6
pytest-rerunfailures==9.1
pytest-soft-assertions==0.1.2
pytest-xdist==2.1.0
requests==2.24.0
selenium==3.141.0
six==1.15.0
softest==1.2.0.0
texttable==1.6.3
toml==0.10.1
urllib3==1.25.10
webdriver-manager==3.2.2
wrapt==1.12.1
xlrd==1.2.0